fix(events): move EMT confirm from per-ticket (guest list) to per-order (sales tab) (#9)
- New POST /api/orders/[id]/confirm-payment — confirms all held e-Transfer
  tickets in an order atomically, sends one receipt + one ticket bundle email
- Extract shared confirmHeldTickets helper into apps/events/src/lib/confirm-payment.ts
- Refactor existing /api/tickets/[id]/confirm-payment to use the shared helper
  and add deprecation warning log
- Sales tab: add 'Confirm e-Transfer' button in Actions column for pending
  e-Transfer orders, with confirmation dialog
- Guest list: remove per-ticket confirm button, replace with muted
  'Pending e-Transfer' badge
